Revert "feat: addition of refactor for root domain on IGS to support main sit…"
This reverts commit 64b31effd6d9f0a13654531b7472fa3bfbe35656.
