+++ src/lib/p5/igsSketch.ts
@@ -26,7 +26,7 @@ export const igsSketch = (p5: any) => {
 	P5Store.set(p5);
 
 	p5.preload = () => {
-		p5.font = p5.loadFont('/slicer/fonts/PlusJakartaSans/VariableFont_wght.ttf');
+		p5.font = p5.loadFont('/fonts/PlusJakartaSans/VariableFont_wght.ttf');
 	};
 
 	p5.setup = () => {
