+++ svelte.config.js
@@ -11,8 +11,7 @@ const config = {
 		// adapter-auto only supports some environments, see https://kit.svelte.dev/docs/adapter-auto for a list.
 		// If your environment is not supported or you settled on a specific environment, switch out the adapter.
 		// See https://kit.svelte.dev/docs/adapters for more information about adapters.
-		adapter: adapter(),
-		paths: { base: '/slicer' }
+		adapter: adapter()
 	}
 };
